Put label and node logics into 2 files in k8snodedecorator pkg
