+++ .gitignore
@@ -2,3 +2,4 @@
 .vscode/
 
 bin/
+.DS_Store
